DFU speed: add callback to SMP BLE transport writes
This provides a signal to the app when writes are occurring on the SMP characteristic.
This gives the app an opportunity to request a BLE configuration to change to speed up the DFU.
It also has the side effect that any activity on the SMP characteristic will signal so
file transfers and other SMP behaviour will send the signal and will be sped up.
The app will be expected to timeout some time after activity has ceased to return to a default speed
for power reduction purposes or other requirements.